Suppressed UserProfile component.
