+++ kansha/app/comp.py
@@ -12,25 +12,27 @@ import cgi
 import sys
 import json
 import urlparse
+from collections import OrderedDict
 
 import webob
 import configobj
 import pkg_resources
 
-from nagare.i18n import _
+from nagare.i18n import _, _L
 from nagare.admin import command
 from nagare.namespaces import xhtml5
 from nagare import component, wsgi, security, config, log, i18n
 
 from kansha import events
 from kansha import exceptions
-from kansha.user import user_profile
+from kansha.menu import MenuEntry
 from kansha.authentication import login
 from kansha import services, notifications
 from kansha.services.search import SearchEngine
 from kansha.user.usermanager import UserManager
 from kansha.board.boardsmanager import BoardsManager
 from kansha.security import SecurityManager, Unauthorized
+from kansha.user.user_profile import get_userform, UserBoards  # !!!!!!!!!!!!!!!
 
 
 def run():
@@ -56,7 +58,21 @@ class Kansha(object):
         self.user_menu = component.Component(None)
         self.content = component.Component(None)
         self.user_manager = UserManager()
-        self.boards_manager = self._services(BoardsManager, self.app_title, self.app_banner, self.theme, card_extensions, self.search_engine)
+        self.boards_manager = self._services(
+            BoardsManager, self.app_title, self.app_banner, self.theme,
+            card_extensions, self.search_engine)
+
+        self.home_menu = OrderedDict()
+        self.selected = 'board'
+
+    def _on_menu_entry(self, id_):
+        """Select a configuration menu entry
+
+        In:
+            - ``id_`` -- the id of the selected menu entry
+        """
+        self.content.becomes(self.home_menu[id_].content)
+        self.selected = id_
 
     def initialization(self):
         """ Initialize Kansha application
@@ -67,8 +83,32 @@ class Kansha(object):
         Return:
          - app initialized
         """
-        self.user_menu = component.Component(security.get_user())
-        if security.get_user() and self.content() is None:
+        user = security.get_user()
+        self.home_menu['boards'] = MenuEntry(
+            _L(u'Boards'),
+            'board',
+            self._services(
+                UserBoards,
+                self.app_title,
+                self.app_banner,
+                self.theme,
+                self.card_extensions,
+                user.data,
+                self.search_engine
+            )
+        )
+        self.home_menu['profile'] = MenuEntry(
+            _L(u'Profile'),
+            'user',
+            self._services(
+                get_userform(
+                    self.app_title, self.app_banner, self.theme, user.data.source
+                ),
+                user.data,
+            )
+        )
+        self.user_menu = component.Component(user)
+        if user and self.content() is None:
             self.select_last_board()
         return self
 
@@ -90,6 +130,7 @@ class Kansha(object):
         board = self.boards_manager.get_by_id(id_)
         if board is not None and not board.archived:
             self._select_board(board)
+            self.selected = 'board'
         else:
             raise exceptions.BoardNotFound()
 
@@ -117,18 +158,7 @@ class Kansha(object):
         if data_board and not data_board.archived and data_board.has_member(user):
             self.select_board(data_board.id)
         else:
-            self.content.becomes(
-                self._services(
-                    user_profile.UserProfile,
-                    self.app_title,
-                    self.app_banner,
-                    self.theme,
-                    self.card_extensions,
-                    user.data,
-                    self.search_engine
-                ),
-                'edit'
-            )
+            self._on_menu_entry('boards')
 
     def handle_event(self, event):
         if event.is_(events.BoardLeft) or event.is_(events.BoardArchived):
